+++ grimmlink.koplugin/_meta.lua
@@ -3,5 +3,5 @@ return {
     name = "grimmlink",
     fullname = _("GrimmLink"),
     description = _("KOReader Companion for Grimmory"),
-    version = "v1.3.1",
+    version = "v1.3.3",
 }